PSHE-453

+++ src/main/java/com/example/Timesheet/com/controller/TimesheetRowController.java
@@ -1,6 +1,7 @@
 package com.example.Timesheet.com.controller;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.CrossOrigin;
@@ -11,6 +12,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.example.Timesheet.com.GlobalVars;
 import com.example.Timesheet.com.dto.TimesheetRowDTO;
 import com.example.Timesheet.com.mapper.TimesheetRowMapper;
 import com.example.Timesheet.com.model.TimesheetRow;
@@ -34,14 +36,12 @@ public class TimesheetRowController {
 	@GetMapping("/timesheetRow/one")
 	public TimesheetRow findAllTimesheetRows(@RequestParam(value="id") int id){
 		
-		return timesheetRowService.getById(id);
+		return timesheetRowService.getById(id).get();
 		
 	}
 	
 	@PostMapping("/timesheetRow")
 	public void post(@RequestBody TimesheetRowDTO timesheetRowDto) {
-		
-
 		TimesheetRow timesheetRow = this.timesheetRowMapper.DTOtoTimesheetRow(timesheetRowDto);
 		
 		this.timesheetRowService.postTimesheetRow(timesheetRow);
@@ -49,12 +49,18 @@ public class TimesheetRowController {
 	}
 	
 	@DeleteMapping("/timesheetRow")
-	public void delete(@RequestBody TimesheetRowDTO timesheetRowDto) {
+	public String delete(@RequestParam(value="id") int id) {
 		
-		TimesheetRow timesheetRow = this.timesheetRowMapper.DTOtoTimesheetRow(timesheetRowDto);
+		Optional<TimesheetRow> optionalRow = this.timesheetRowService.getById(id);
 		
+		if(optionalRow.isPresent() == false) {
+			return GlobalVars.TimesheetRowIdNotFound;
+		}
+		
+		TimesheetRow timesheetRow = optionalRow.get();
 		this.timesheetRowService.deleteTimesheetRow(timesheetRow);
 		
+		return GlobalVars.TimesheetRowDeleteSuccessful;
 	}
 
 }
